src-python/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
pynput==1.7.6
PyObjC==10.1
pyperclip==1.8.2
Pillow==10.2.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pyinstaller==6.3.0
requests==2.31.0
matplotlib==3.8.0
numpy==1.26.0
python-multipart==0.0.6
